Feat: 친구 관리 화면 구현 #17
